Исправляет замечания. Меняет отрисовку фильтра. Ошибки в консоли
